requirements.txt
flask
flask-cors
playwright
google-generativeai